% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballantine.R
\name{ballantine}
\alias{ballantine}
\title{Creates ballantine graphs}
\usage{
ballantine(data, dep, ind_list, round = 2, leftover = T)
}
\arguments{
\item{data}{dataframe object to be used}

\item{dep}{dependent variable to be regressed on. Must be a character type.
Column must exist in \code{data}.}

\item{ind_list}{character vector of independent variables to regress
dependent variable on. Max length of 3. Columns must exist in \code{data}.}

\item{round}{number of decimal points to round values to. Defaults to 2.}

\item{leftover}{logical indicating whether the leftover/unexplained variance
should be included outside of the main ballantine graph. Defaults to
TRUE.}
}
\description{
Based on a dataset, dependent variable, and up to 3 independent variables,
creates a ballantine plot displaying the amount of shared and unique
variance is explained by each independent variable to the dependent variable.

Produces a ggplot object. Best used as a visualization shorthand.
}
\examples{
#one independent variable
ballantine(mtcars, "hp", "disp")

#two independent variables
ballantine(mtcars, "hp", c("disp", "wt"))

#three independent variables
ballantine(mtcars, "hp", c("disp", "wt", "qsec"))
}
